// packages/database/src/mongoose/interfaces/iThresholdCreateInput.ts
// THIS CODE WAS AUTOMATICALLY GENERATED
import {databaseTypes} from 'types';
import {Types as mongooseTypes} from 'mongoose';

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface IThresholdCreateInput extends Omit<databaseTypes.IThreshold, '_id' | 'createdAt' | 'updatedAt'> {}